LayoutTests/platform/mac/accessibility/select-text.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
<title>Select Text</title>
</head>
<body>

<p contenteditable="true" id="text">The quick brown fox <span id="target">jumps</span> over the lazy dog.</p>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests the ability to select and replace text with respect to selection.");
    
    function selectedText() {
        return window.getSelection().toString();
    }
    
    function selectElementText(element) {
        var range = document.createRange();
        range.selectNodeContents(element);
        
        var selection = window.getSelection();
        selection.removeAllRanges();
        selection.addRange(range);
    }
    
    if (window.accessibilityController) {
        var text = accessibilityController.accessibleElementById("text");
        var result = 0;
        var selection = 0;
        var target = document.getElementById("target");
        
        // Select text after selection (single search string).
        selectElementText(target);
        result = text.selectTextWithCriteria("AXSelectTextAmbiguityResolutionClosestAfterSelection", "lazy");
        shouldBe("result", "'lazy'");
        selection = selectedText();
        shouldBe("selection", "'lazy'");
        
        // Select text after selection (multiple search strings).
        selectElementText(target);
        result = text.selectTextWithCriteria("AXSelectTextAmbiguityResolutionClosestAfterSelection", ["lazy", "over"]);
        shouldBe("result", "'over'");
        selection = selectedText();
        shouldBe("selection", "'over'");
        
        // Select text before selection (single search string).
        selectElementText(target);
        result = text.selectTextWithCriteria("AXSelectTextAmbiguityResolutionClosestBeforeSelection", "quick");
        shouldBe("result", "'quick'");
        selection = selectedText();
        shouldBe("selection", "'quick'");
        
        // Select text before selection (multiple search strings).
        selectElementText(target);
        result = text.selectTextWithCriteria("AXSelectTextAmbiguityResolutionClosestBeforeSelection", ["quick", "fox"]);
        shouldBe("result", "'fox'");
        selection = selectedText();
        shouldBe("selection", "'fox'");
        
        // Select text closest to selection (single search string).
        selectElementText(target);
        result = text.selectTextWithCriteria("AXSelectTextAmbiguityResolutionClosestToSelection", "dog");
        shouldBe("result", "'dog'");
        selection = selectedText();
        shouldBe("selection", "'dog'");
        
        // Select text closest to selection (multiple search strings).
        selectElementText(target);
        result = text.selectTextWithCriteria("AXSelectTextAmbiguityResolutionClosestToSelection", ["dog", "fox"]);
        shouldBe("result", "'fox'");
        selection = selectedText();
        shouldBe("selection", "'fox'");
        
        // Select and replace text closest to selection (single search string).
        selectElementText(target);
        result = text.selectTextWithCriteria("AXSelectTextAmbiguityResolutionClosestToSelection", "quick", "slow");
        shouldBe("result", "'slow'");
        selection = selectedText();
        shouldBe("selection", "'slow'");
        
        // Select and replace text closest to selection (multiple search strings).
        selectElementText(target);
        result = text.selectTextWithCriteria("AXSelectTextAmbiguityResolutionClosestToSelection", ["dog", "fox"], "cat");
        shouldBe("result", "'cat'");
        selection = selectedText();
        shouldBe("selection", "'cat'");

        // Select brown and capitalize it.
        selectElementText(target);
        result = text.selectTextWithCriteria("AXSelectTextAmbiguityResolutionClosestToSelection", "brown", null, "AXSelectTextActivityFindAndCapitalize");
        shouldBe("result", "'Brown'");
        selection = selectedText();
        shouldBe("selection", "'Brown'");

        // Select Brown and lowercase it.
        selectElementText(target);
        result = text.selectTextWithCriteria("AXSelectTextAmbiguityResolutionClosestToSelection", "Brown", null, "AXSelectTextActivityFindAndLowercase");
        shouldBe("result", "'brown'");
        selection = selectedText();
        shouldBe("selection", "'brown'");

        // Select brown and uppercase it.
        selectElementText(target);
        result = text.selectTextWithCriteria("AXSelectTextAmbiguityResolutionClosestToSelection", "brown", null, "AXSelectTextActivityFindAndUppercase");
        shouldBe("result", "'BROWN'");
        selection = selectedText();
        shouldBe("selection", "'BROWN'");

    }
</script>

<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
